Intermediary Server/requirements.txt
pycrypto
pycryptodome
flask
python-dotenv==0.19.0
gunicorn
requests
